Mesclagem de mapLoad com Map e limpeza no codigo
