+++ src/br/unicamp/mc322/lab10/projeto/Coordinate.java
@@ -15,17 +15,13 @@ public class Coordinate {
 		this.y = y;
 		onMap = true;
 	}
-	
-	public void setOnInventory() {
-		onMap = false;
+
+	public boolean equals(Coordinate position) {
+		if(x == position.getX() && y == position.getY())
+			return true;
+		return false;
 	}
-	
-	public void setOnMap(int x, int y) {
-		setX(x);
-		setY(y);
-		onMap = true;
-	}
-	
+
 	public int getX(){
 		if (!onMap)
 			return -1;
@@ -37,28 +33,7 @@ public class Coordinate {
 			return -1;
 		return y;
 	}
-	
-	public void setX(int x){
-		this.x = x;
-	}
-	
-	public void setY(int y){
-		this.y = y;
-	}
-	
-	public Boolean equals(Coordinate position) {
-		if(x == position.getX() && y == position.getY())
-			return true;
-		return false;
-	}
-	
-	private int positivize(int number) {
-		/* Torna numero dado em valor positivo */
-		if(number < 0)
-			return number*-1;
-		return number;
-	}
-	
+
 	public int measureDistance(Coordinate position) {
 		/* Retorna a soma das distancias em X e Y da coordenada passada
 		 * com a do objeto atual */
@@ -76,4 +51,30 @@ public class Coordinate {
 			return "Objeto não esta no mapa";
 		return "("+x+","+y+")";
 	}
+
+	public void setX(int x){
+		this.x = x;
+	}
+	
+	public void setY(int y){
+		this.y = y;
+	}
+	
+	public void setOnInventory() {
+		onMap = false;
+	}
+	
+	public void setOnMap(int x, int y) {
+		setX(x);
+		setY(y);
+		onMap = true;
+	}
+	
+	private int positivize(int number) {
+		/* Torna numero dado em valor positivo */
+		if(number < 0)
+			return number*-1;
+		return number;
+	}
+
 }
