+++ src/br/unicamp/mc322/lab10/projeto/mapObjects/objects/mapItems/forniture/fornitureTypes/Chest.java
@@ -14,20 +14,23 @@ public class Chest extends Forniture{
 	private static final char SPRITE = 'C';
 	private static final GameTypeObjects ID = GameTypeObjects.CHEST;
 	private static final String NAME = "Chest";
-	private CanCarry[] content;
+	private CanCarry content;
 	private Boolean open = false;
 	
-	public Chest(CanCarry[] content,Coordinate position) {
+	public Chest(Coordinate position) {
 		super(NAME,new Sprite(SPRITE),ID, position);
-		this.content = content;
 	}
 	
-	public CanCarry[] getItems() {
+	public CanCarry getItems() {
 		if(!open) {
 			open = true;
 			return content;
 		}
 		
 		return null;
+	}
+	
+	public void setContent(CanCarry content) {
+		this.content = content;
 	}
 }
